Just block scrolling, instead of entire applet
